requirements.base.txt
# ---- Core / Data ----
yfinance>=0.2.48,<0.3
pandas>=2.2,<3.0
numpy>=1.26,<3.0
scikit-learn>=1.5,<1.7
pyyaml>=6.0.2,<7.0
joblib>=1.4,<2.0

# ---- Viz / UI / Logging ----
matplotlib>=3.8,<3.10
tensorboard>=2.14,<3.0
tqdm>=4.66,<5.0
streamlit>=1.31,<2.0
